Added assignment operator
This fixes a bug when subdividing the mesh: The old mesh is discarded
and the new mesh should replace the old one. Since pointers are used and
because the destructor of the old mesh is called, the pointer
_addresses_ will be copied, but the pointer _data_ will be deleted. This
would yield errors on access.

Solution consists of adding assignment operators that copy the pointer
_data_ and create new pointers.