Added "replace_with" function
Because of the pointers, replacing a mesh with another one is not
straightforward. Hence,  a function "replace_with" has been added which
simply copies all data from one mesh to another and _clears_ the other
mesh's data structures afterwards, which would result in memory leaks if
the mesh had not been copied.