Added sorting function for vertices
Added a function that allows sorting the vertices of a face in
counterclockwise order, thereby allowing a more parametric approach to
subdivision algorithms.

This _should_ reduce computational complexity because the order of
vertices needs to be determined once per face (in O(k), where k is the
number of edges of the faces). In order to create all face vertices for
the Doo-Sabin algorithm, for example, it is sufficient to _shift_ the
vector of vertices by one.

This needs to be implemented.