Added segmentation algorithms
Unusable and hard-coded for now. General idea is to define regions of
planarity among meshes and store these regions separately. This shall be
used for removing planar parts of an arbitrary mesh.

Ideas so far:

* Use RMS curvature
* Identify "almost" planar vertices
* _Try_ to identify a region containing these vertices
* Make this region as large as possible

This requires severe changes in the mesh class. Namely, it should be
possible to _copy_ parts of the mesh -- not a great idea as the vertex
indices might change.